+++ common/util.py
@@ -1104,3 +1104,9 @@ def check_notifications():
     log_action(consts.MSG_NOTIFICATION_NOT_FOUND)
 
   time.sleep(0.2)
+
+def open_map(x1, y1, x2, y2):
+  move_click(835, 705, 0.2)
+  move_click(865, 315, 0.2)
+  move_click(x1, y1, 0.2)
+  move_click(x2, y2, 0.2)